==> docs/tutorial/01-introduction.ipynb <==
Introduction
===

This tutorial describes the process of analyzing bibliographic information using **techMiner**. 

The library can be installed using:

```
!pip3 install techminer
```

directly in a Jupyter notebook.

For creating a worldmap plot the computer must have installed [GeoPandas](http://geopandas.org/install.html).

For the appropiate use of the library, the user must have a basic knowledge of:

* Python.

* Matplotlib.

* Pandas.

* Seaborn (optional for some plots).

* Altair (optional for some plots).

Please refer to the documentation of these tools in case of any doubt.
