==> docs/tutorial/03-step-02-file-merging.ipynb <==
# Step 2: Data merging and storing on json format

From previous step, there are three files with the data in the current directory:

!ls -1 *.csv

First, the files are loaded as dataframes.

import pandas as pd

citations = pd.read_csv("demo-citations.csv")
keywords = pd.read_csv("demo-abstracts.csv")
references = pd.read_csv("demo-references.csv")

Second, column names are listed and reviewed, looking for unwanted columns.

citations.columns

keywords.columns

references.columns

For all dataframes, the `'Link'` column can be removed. Also, the column `'Authors with affiliations'` in `citations` can be removed.

citations = citations.drop(columns=["Link", "Authors with affiliations"])
keywords = keywords.drop(columns="Link")
references = references.drop(columns="Link")

Third, the dataframes are concated.

df = pd.concat([citations, keywords, references], axis=1)

Finally, the dataframe is saved to disk in json format for further processing.

df.to_json("step-02.json", orient="records", lines=True)

In the next part of this tutorial, strategies for selecting records are discussed.
