==> docs/tutorial/05-step-04-data-deleting.ipynb <==
# Step 4: Deleting of records

import pandas as pd
from techminer import RecordsDataFrame

rdf = RecordsDataFrame(pd.read_json("step-03.json", orient="records", lines=True))

## Deleting of abstracts of proceedings and workshops 

from techminer import Keywords

conf = Keywords()
conf.add_keywords(["Conference", "Proceeding", "Workshop", "Congress"])
rdf["CONF"] = rdf["Title"].map(lambda x: True if x in conf else False)

for title in rdf[rdf["CONF"]]["Title"]:
    print(title)

print("Num. records before = ", len(rdf))
rdf = rdf[rdf["CONF"].map(lambda x: False if x is True else True)]
print("Num. records after = ", len(rdf))

## Deleting of documents with the same title

from techminer import remove_duplicate_records

print("Num. records before = ", len(rdf))
rdf = remove_duplicate_records(rdf, "Title", match_type="fingerprint")
print("Num. records after = ", len(rdf))

rdf.to_json("step-04.json", orient="records", lines=True)
