==> docs/tutorial/06-step-05-keywords-cleanup.ipynb <==
# Step 5: Keywords cleanup

import pandas as pd
from techminer import RecordsDataFrame

rdf = RecordsDataFrame(pd.read_json("step-04.json", orient="records", lines=True))

## Keywords with equal number of words

from techminer import Thesaurus, text_clustering

#
# Search of keywords with the same root
#
th = text_clustering(rdf['keywords'], sep=';', transformer=lambda x: x.lower())
with open('thesaurus-1.json', 'w') as f:
    f.write(th.__repr__())
!head -n 30 thesaurus-1.json

#
#  Number of strings differing in keywords
#
print(
    len(
        set([w.strip() for x in rdf["keywords"] if x is not None for w in x.split(";")])
    )
)

#
# Reading of keywords edited by the analyst
#
import json

with open("thesaurus-1-edited.json", "r") as f:
    dictionary = json.loads(" ".join(f.readlines()))

##
## Limpieza
##
from techminer import Thesaurus

th = Thesaurus(dictionary, ignore_case=False, full_match=True, use_re=False)
rdf["keywords (cleaned)"] = rdf["keywords"].map(lambda x: th.apply(x, sep=";"))
rdf["keywords (cleaned)"] = rdf["keywords (cleaned)"].map(
    lambda x: ";".join(set([w.strip() for w in x.split(";")]))
)
rdf["keywords (cleaned)"] = rdf["keywords (cleaned)"].map(
    lambda x: x if x != "" else None
)

#
# Number unique of strings
#
print(
    len(
        set(
            [
                w.strip()
                for x in rdf["keywords (cleaned)"]
                if x is not None
                for w in x.split(";")
            ]
        )
    )
)

## Substrings clustering

#
# Substrings in keywords
#
from techminer import text_nesting
tn = text_nesting(rdf['keywords (cleaned)'], sep=';', max_distance=1, transformer=lambda x: x.lower())
with open('thesaurus-2.json', 'w') as f:
    f.write(tn.__repr__())
!head -n 60 thesaurus-2.json

!head -n 60 thesaurus-2-edited.json

with open("thesaurus-2-edited.json", "r") as f:
    dictionary = json.loads(" ".join(f.readlines()))
th = Thesaurus(dictionary, ignore_case=False, full_match=True, use_re=False)

rdf["keywords (cleaned)"] = rdf["keywords (cleaned)"].map(
    lambda x: th.apply(x, sep=";")
)
rdf["keywords (cleaned)"] = rdf["keywords (cleaned)"].map(
    lambda x: ";".join(set([w.strip() for w in x.split(";")]))
)
rdf["keywords (cleaned)"] = rdf["keywords (cleaned)"].map(
    lambda x: x if x != "" else None
)

#
# Number of unique keywords
#
print(
    len(
        set(
            [
                w.strip()
                for x in rdf["keywords (cleaned)"]
                if x is not None
                for w in x.split(";")
            ]
        )
    )
)

#
# Review
#
from techminer import display_records

display_records(rdf[["Title", "keywords (cleaned)"]].head(10))

## Deleting based on keywords or other text

from techminer import display_records

display_records(rdf[["Title", "keywords (cleaned)"]].head(10))

from techminer.keywords import Keywords

kyw = Keywords()
kyw.add_keywords(["Vacuum", "market data"])
kyw

"vacuum" in kyw

idx = rdf["keywords (cleaned)"].map(lambda x: not kyw.common(x, sep=";"))

idx[0:20]

print("Records before = ", len(rdf))
rdf = rdf[idx]
print("Records after = ", len(rdf))

rdf.to_json("step-05.json", orient="records", lines=True)
