==> docs/tutorial/08-step-07-references.ipynb <==
# Step 7: Numbering of references

import pandas as pd
from techminer import RecordsDataFrame

rdf = RecordsDataFrame(pd.read_json("step-06.json", orient="records", lines=True))

#
# A column for identify each record is added to the dataframe.
#
rdf = RecordsDataFrame(rdf).generate_ID()

rdf.to_json("step-07.json", orient="records", lines=True)
