==> docs/tutorial/09-step-08-direct-queries.ipynb <==
# Step 7: Direct queries over data

import pandas as pd
from techminer import RecordsDataFrame

rdf = RecordsDataFrame(pd.read_json("step-07.json", orient="records", lines=True))

Some directec queries can be made directly over the dataframe without recurring to special functions or code.

#
# Number of records in the dataframe
#
len(rdf)

#
# Number of sources
#
rdf.num_of_sources

#
# Top N most cited documents
#
rdf.most_cited_documents(top_n=10)
