==> docs/tutorial/15-terms-by-terms-by-year.ipynb <==
# Analysis: Terms by terms by year

import pandas as pd
from techminer import RecordsDataFrame

rdf = RecordsDataFrame(
    pd.read_json(
        'step-07.json', 
        orient='records', 
        lines=True))

## Authors vs Source Title

rdf.terms_by_terms_by_year(
    column_r='Authors', 
    column_c='Source title', 
    sep_r=',', 
    sep_c=None)
